Fix author dc current for avatar in fse - backend and frontend
